Add missing header for non-debug builds
The display() method for SYCL's cl::sycl::detail::display_vector struct
uses std::cout, but the <iostream> header is only included for debugging
builds. This causes non-debug builds to break due to use of the
undefined reference. This patch removes the TRISYCL_DEBUG conditional
guard around the inclusion of the iostream header.