Major refactoring of asynchronous execution towards supporting latest 1.2 specification
The asynchronous execution mode was never completed with the command
groups as defined in the provisional specification.

This is a first attempt for supporting asynchronous execution with the
command group handlers as defined in the final 1.2 SYCL specification.

The tests in synchronous execution mode mostly pass for now, besides a
dead lock in accessor/demo_parallel_matrix_add.

There are still a lot of failures in asynchronous mode.