In synchronous mode, executes also all the dependency verification code
Fix the deadlock on accessor/demo_parallel_matrix_add.cpp test.
So now all the dependency checking is done, even if the kernels are
executed in order.
The nice side effect is that it will lock waiting for some buffer that
will never be produced instead of having undefined behavior for some
applications that cannot be executed with kernels in order.