Fix 3 issues in a test example failing in asynchronous mode
Avoid 2 kernels in the same command group Use 2 submissions to the
queue instead.
Accessor needs to be captured by value and not by reference since the
original may no longer exist when executing in another thread.
Since the buffer is a global object, add an host accessor to access
its content.