Fix several issues in pipe reservation
Fix size computation.
Use move construction of reservation station and fix the move
constructor to avoid commiting in the destructor of the moved object.
Add mutable state for survive to lambda expressions.
Verify that the reservation is in a good state before using it.
Add debuging code.
Improve example to use work-group loops.