Add variadic set_args() to set all kernel arguments at once in OpenCL kernel interoperability mode
This is discussed in
https://cvs.khronos.org/bugzilla/show_bug.cgi?id=15978

In the current SYCL specification, when using plain OpenCL kernels with the interoperability mode, it is only possible to set arguments one at a time with handler::set_arg().

This is verbose and cumbersome, whereas Khronos cl2.hpp and Boost.Compute provide ways to do this based on variadic templates.

So add a

  template <typename... Ts> void set_args(Ts&&... args)

for this.

I think this should be also back-ported to SYCL 1.2.1.

Question for an OpenCL lawyer:
Should we verify that the number of arguments matches the kernel arity?
I have the feeling that we should not because we may imagine a kernel
with some trailing unused arguments and in that case it is useless to
set them.