Add travis-ci support for triSYCL
Every commit will trigger two travis-ci builds, one with gcc-6
and the other with clang-3.9. The build script  will compile and
run the triSYCL tests with cmake and ctest failing if either commands
fail.

The travis-ci build use docker to run on an ubuntu 16.10 environment,
neither of the build use OpenCL for now and only the gcc-6 build
uses OpenMP.