buffer: fix host context constructor.
This fixes two bugs in the buffer destructor.

The first is the destructor tries to instantiate a context
object but since this is a destructor that could be getting
call on program exit, the context object is possibly
already destroyed. (valgrind showed this). This takes a reference
to the host context on construction and uses that.

The second is that the default host constructor uses default
device selection, this forces it to use the host device which
is returned by the default device constructor.