vec: initial swizzle implementation (v5)
This isn't at all spec compliant mainly around the __swizzled_vec__
type and the rules it's meant to follow.

But it adds the APIs and some tests, and is probably half way to
what the proper answer looks like.

This adds src/gen.c which is the C source used to generate the swizzle header files.
I don't think these files need to change that much, so I haven't made it run, and just
checked in the result, we can revisit this later if required.

v2: remove do_swizzle and just have swizzle be in the base class, and call directly
to it (Ronan)
v3: fix hi for 8 vec, fix missing checks in test, replace recursive pack param (Ronan)
v4: don't use auto in tests.
v5: drop second macro for swizzle defs

realiseed we only need one macro