Refactored `parallel_for_workitem()` implementation
I replaced the older OpenMP modulo/divison loop with two suggested newer
implementations that use OpenMP parallel collapses and OpenMP SIMD
operations to speed things up. There are now two separate
implementations that use OpenMP. One  that is active when
TRISYCL_NO_BARRIER is defined and the other when it's not defined.
The implementation when TRISYCL_NO_BARRIER is defined is a little
less strict and makes use of SIMD operations.

I made use of constexpr rather than template specializations to
specialise the OpenMP loops for each dimension.

Modifying this however makes the results non-deterministic enough that
printing output and regex'ng for it is no longer a sufficient testing
mechanism for hierarchical.cpp and hierarchical_new.cpp (previously
TRISYCL_NO_BARRIER allowed it to default to a non-OpenMP implementation).
A new implementation for these tests has been created that assigns values
to buffers and checks the output is correct.

Related to triSYCL issue: https://github.com/triSYCL/triSYCL/issues/43