Adding Some Vec Math Functions
These were the functions required to get SYCL SLAMBench functioning, 
they're intended for use with the Vec class. They don't function exactly 
like Codeplays implementation at the moment, these are a little stricter 
with the specification (Codeplay lets you pass in vectors for minval and 
maxval on clamp for instance, which doesn't appear to be in the SYCL 
spec at the moment. It could be a handy addition though.).