WIP parallel_for ND_RANGE_KERNEL review update
A patch aimed at addressing all of the prior change requests in some 
way. Also adding two new device_compiler tests and a slight modification 
to id.hpp.

Modifications:

id.hpp: Added dimensionality as a public variable similarly to range.hpp 
as there appears to be no other simple way to introspect id's 
dimensionallity without type deduction at the moment. Unsure if 
dimensionality was intentionally left out of id.hpp and if so it can be 
removed (I'd have to readjust opencl_spir_helpers.hpp shouldn't be an 
issue however).

environment.rst, architecture.rst and macros.rst: Documentation 
updates/corrections

hanler.hpp: tidied in a previous commit and fixed some comments with 
this commit.

opencl_spir_req.h: Added ifndef/define even if its not needed it 
probably doesn't hurt to have it! Also removed the top-level data types 
and changed the email/name in the documentation to my own.

opencl_spir_helpers.hpp: Changing the email/name. Removing old 
gen_spir_id/gen_spir_range and enum class then replacing with 
fill_id_or_range (is there a better name?) which takes a function and 
invokes it for each dimension of the id or range passing the dimension 
to the function. Modified the other functions accordingly and added an 
else if constexpr rather than two seperate if constexpr's.

parallelism.hpp: Addition of new template function capture_arg_v which 
captures the argument of a simple 1 argument passed in lambda and 
returns a instance of the captured argument, replaces what mem_fn was 
previously doing in parallel_for as the member functions we rely on will 
be deprecated soon (which weirdly looks like mem_fn is left as a 
redundant implementation of std::invoke). I also added comments 
describing its usage and why it's used in this context so that it's 
possible to read and understand the reasoning a little better.

Makefile: Fixing indenting of several things I've added that are 
enclosed in ifdef's so that they suit the formatting that existed 
before. Adding the XILINX_SPECIFIC_PASS variable and appending 
-reqd-workgroup-size-t to it as suggested. Also removing a commented out 
piece of code that I think is no longer required.

Added:

parallel_for_ND_range.cpp & parallel_for_ranges.cpp: Two new tests the 
latter tests that range data is correct on a basic level and the former 
checks that ND range kernels are executing in more than 1D and tests 
get_linear_id and the ranges further.