Additional Device Compiler Tests
Several new device compiler tests that work with POCL when they're
compiled using a newer version of the triSYCL device compiler (after the
commit: [OpenCL] Add generic AS to 'this' pointer).

However, the buffer_of_objects_members.cpp test works with the
sycl/release_70/master branch of the triSYCL device compiler.

The buffer of objects tests are related to:
https://github.com/triSYCL/triSYCL/issues/212

The ctor_and_dtor tests are related to:
https://github.com/triSYCL/triSYCL/issues/213