Empty base class fix to small_array
Related to: https://github.com/triSYCL/triSYCL/issues/230

small_array's empty base class optimization was not working correctly 
causing the vec class (and possibly others) to bloat by a few extra 
bytes. A large problem for cl::sycl::vec as alignas gets invoked so that 
it's correctly aligned to the data it contains, however in the case 
where its larger by a few bytes alignas was rounding it up to double the 
size it should be. The fix is basically chaining the boost operation 
classes that were causing the EBCO issue (chaining boost operation 
classes is an inbuilt boost workaround for this very issue it seems, 
link in the comments of small_array.hpp).

Added a new test to check that vec's are appropriately sized and 
contigious (vecmemlayoutalign.cpp) and modified the CMakeList.txt to 
include it in the test suite.

I also tweaked the vec.hpp classes get_size method as it was incorrect 
for vec's of 3 elements (special case for it to be treated as sizeof(T) 
* 4 in SYCL Spec 1.2.1 Section 4.10.2.6). Also fixing the vecacc.cpp 
test to reflect this change.