set_final_data overload fix
Fix to something I broke in a prior patch that was causing build/compile 
errors for tests in the SYCLParallelSTL.

Didn't account for the need to remove the reference on the rvalue 
referened iterator when passed in, so the iterator_traits class could 
not find the value_type member!

One of those little hard to catch move/reference semantic errors.

Also some minor spelling fixes.