Rename cl::sycl:: to trisycl:: and add a cl::sycl:: compatibility mode
This is the start of implementing
https://github.com/triSYCL/triSYCL/issues/228
and the ability to have several SYCL implementation to coexist and
having layered implementations.

First looked at clang-rename to rename cl::sycl to trisycl but it does
not work because it is more a renaming + namespace fusion. But since
we are using C++17 namespaces now, a simple regex solution works. :-)

  find . -type f -name '*.hpp' -exec sed -i 's/cl::sycl/trisycl/g' {} +

Some manual work is required to qualify better some names because
there are more namespace ambiguity, such as trisycl,
trisycl::vendor::trisycl, trisycl::detail and
trisycl::vendor::trisycl::scop::detail