Rename trisycl directory to triSYCL and trisycl.hpp to triSYCL/sycl.hpp
This limit the conflicts of includes at the top level.

introduce triSYCL in the trisycl:: namespace and do not use cl::sycl::
or sycl::, leaving room for other implementations to coexist.

Also, while it is an extension, using triSYCL/sycl.hpp instead of
trisycl.hpp remains more in the spirit of the SYCL 1.2.1 specification
along CL/sycl.hpp.