Rename sycl.hpp to SYCL/sycl.hpp
This limit the conflicts of includes at the top level.

introduce triSYCL in the sycl:: namespace and do not use cl::sycl::
leaving room for other implementations to coexist in cl::sycl:: or
just provide a short cut saving 4 letters.

Also, while it is an extension, using SYCL/sycl.hpp instead of
sycl.hpp remains more in the spirit of the SYCL 1.2.1 specification
along CL/sycl.hpp.