[AIE_HW] Tweak Webcam, remove load etc (now inside API), uncomment kernel
This should be ready for one step compilation for the moment, the kernel 
and host should compile. And I've stuck the array into an #ifdef for 
some quick two (minimum layout) AIE binary compilation.

The load etc for tiles has been moved into the API for now, hopefully it 
is still functionally equivelant but we will find out if it doesn't work 
and tweak as neccessary!