[AIE_HW] Supports multiple image offloads, only works for 400 tile arrays for now
Added new test templated_noise.cpp, yet to test it, but the intent of it 
is to check the variability of the kernels via constexpr/metaprogramming 
(aim to test once it's compiled, takes a significant time to compile 
larger examples via chess, 5 hours, need a lot of optimizations in place 
to root out unrequired/duplicate kernel compiles and to parallelize the 
script)

sobel_north.cpp alteration to no longer use the #ifdef, we only support 
compilation of BIG examples without changing the device runtime for 
now... double edged sword till optimized/fixed

The runtime only really supports 400 tile arrays for now it seems, seems 
to be a disconnect between the SYCL ACAP runtime and the X AIE engine 
runtime I think. I believe it tries to load the binaries on to the SHIM 
for certain sizes as all the SHIM tiles are in the same AI Engine 
array.. need to look into it.