Invert prerun condition so the default is for a tile to always execute
Returning a postiive from postrun will no longer cancel a tile from 
executing, a user now has to return 0, for a false condition. The main 
reason for this is easy compatability with all of the existing host 
tests that do not define a postrun/prerun and to allow programs the 
flexibility of not defining these functions if they do not want to.