Improve debug tracing class
The trisycl::detail::debug CRTP was using introspection of the native
type to know whether the type was copy and move constructible before
defining such a constructor to avoid preventing the generation of the
default assignment operator.
But actually this is an unsolvable question because of the
self-reference essence of the CRTP.
It was still working with G++ but no longer working starting with
Clang++-11.
So, instead, do not use std::enable_if or concepts but just define all
the operators.
A nice side effect is that now copy and move assignments can be traced
too.