Add action to launch an ssh terminal in GitHub Action
This is for debugging purpose whew required.

Allows you to run this workflow manually from the Actions tab by
selection CMake and then "Run workflow" menu on the right branch
and clicking on "launch_tmate_terminal_for_debug".
Unfortunately this works only for the default branch.
So you can either
- change as the default one default the branch of the PR on the
  GitHub repository owning the PR and launching in Actions tab;
- or edit directly the step below which runs tmate and push to the
  PR, ignoring the manual workflow launch.