Add compilation timeout for partial compilation to ccache
The problem is when the CI compilation time does not fit into the
default GitHub limits, the job is killed and ccache action cannot even
save its compilation cache for next CI execution.
So add an optional timeout usable from the manual launch to stop the
compilation before the deadline so ccache can save the partial compilation.